Cleaned up the savegame manager XML code. Fixed the worldmap so it pulls worldmap information from the XML file instead of hard-coding it. More general cleanup to get rid of some compiler warnings. Made the exit visible on the hellmouth14-grumbel.pingus level. Some of the leveldots are inaccessible in volcano still.

+++ src/worldmap/graph.hxx
@@ -106,7 +106,7 @@ public:
   {
     Edge<EdgeType> new_edge (data, a, b, cost);
     edges.push_back (new_edge);
-    resolve_node (a).next.push_back (edges.size ()-1);
+    resolve_node(a).next.push_back ((int)edges.size()-1);
     return EdgeId (edges.size ()-1);
   }
 
@@ -175,7 +175,7 @@ public:
   }
 
   int max_node_handler_value () {
-    return nodes.size ();
+    return (int)nodes.size ();
   }
 };
